More apps added. See Roadmap.

+++ Applications/NoteBook/3_install.sh
@@ -9,10 +9,19 @@ else
 	. ${_PWD}/desc.sh
 fi
 
-cd ../../../Frameworks/${REPO}/${PATH_SRC} || exit 1
+cd ../../../Applications/${REPO}/${PATH_SRC} || exit 1
+
+if ! [ -f ${APPNAME}.app/${APPNAME} ];then
+	printf "\nNo app ${APPNAME}.app found to install:\nDid You build the app \nwithout any error message ?\n\n"
+	exit 1
+fi
 
 printf "\nYour password maybe required to accomplish the installation:\n"
 sudo -E make install ${INSTALL_ARGS}
+
+if [ -n "${RUN_AT_END}" ];then
+	openapp ${APPNAME}
+fi
 
 make clean
